refactor: move events under session
