+++ src/bidiMapper/domains/context/CdpTarget.ts
@@ -20,13 +20,13 @@ import type {Protocol} from 'devtools-protocol';
 import type {ICdpClient} from '../../../cdp/CdpClient.js';
 import {Deferred} from '../../../utils/Deferred.js';
 import type {Result} from '../../../utils/result.js';
-import type {EventManager} from '../events/EventManager.js';
 import {LogManager} from '../log/LogManager.js';
 import {NetworkManager} from '../network/NetworkManager.js';
 import type {NetworkStorage} from '../network/NetworkStorage.js';
 import type {ChannelProxy} from '../script/ChannelProxy.js';
 import type {PreloadScriptStorage} from '../script/PreloadScriptStorage.js';
 import type {RealmStorage} from '../script/RealmStorage.js';
+import type {EventManager} from '../session/EventManager.js';
 
 export class CdpTarget {
   readonly #targetId: Protocol.Target.TargetID;
