+++ src/bidiMapper/domains/log/LogManager.ts
@@ -18,9 +18,9 @@ import type {Protocol} from 'devtools-protocol';
 
 import {ChromiumBidi, Log, Script} from '../../../protocol/protocol.js';
 import type {CdpTarget} from '../context/CdpTarget.js';
-import type {EventManager} from '../events/EventManager.js';
 import type {Realm} from '../script/Realm.js';
 import type {RealmStorage} from '../script/RealmStorage.js';
+import type {EventManager} from '../session/EventManager.js';
 
 import {getRemoteValuesText} from './logHelper.js';
