+++ src/bidiMapper/domains/session/SubscriptionManager.spec.ts
