+++ src/bidiMapper/domains/session/SubscriptionManager.ts
