+++ src/bidiMapper/domains/session/events.ts
